+++ config/src/terminal.rs
@@ -82,6 +82,10 @@ impl wezterm_term::TerminalConfiguration for TermConfig {
         self.configuration().enable_title_reporting
     }
 
+    fn enable_modify_other_keys(&self) -> bool {
+        self.configuration().enable_modify_other_keys
+    }
+
     fn enable_kitty_keyboard(&self) -> bool {
         self.configuration().enable_kitty_keyboard
     }